Refactor debug view into modular tabs
