+++ components/debug/tabs/MapDataFullTab.tsx
@@ -1,0 +1,16 @@
+import DebugSection from '../DebugSection';
+import type { MapData } from '../../../types';
+
+interface MapDataFullTabProps {
+  readonly mapData: MapData;
+}
+
+const MapDataFullTab = ({ mapData }: MapDataFullTabProps) => (
+  <DebugSection
+    content={mapData}
+    maxHeightClass="max-h-[70vh]"
+    title="Current Map Data (Full)"
+  />
+);
+
+export default MapDataFullTab;
