+++ components/debug/tabs/ThemeHistoryTab.tsx
@@ -1,0 +1,16 @@
+import DebugSection from '../DebugSection';
+import type { ThemeHistoryState } from '../../../types';
+
+interface ThemeHistoryTabProps {
+  readonly themeHistory: ThemeHistoryState;
+}
+
+const ThemeHistoryTab = ({ themeHistory }: ThemeHistoryTabProps) => (
+  <DebugSection
+    content={themeHistory}
+    maxHeightClass="max-h-[70vh]"
+    title="Current Theme History"
+  />
+);
+
+export default ThemeHistoryTab;
