+++ components/debug/tabs/index.ts
@@ -1,0 +1,13 @@
+export { default as CharactersTab } from './CharactersTab';
+export { default as DialogueAITab } from './DialogueAITab';
+export { default as GameLogTab } from './GameLogTab';
+export { default as GameStateTab } from './GameStateTab';
+export { default as InventoryAITab } from './InventoryAITab';
+export { default as InventoryTab } from './InventoryTab';
+export { default as MainAITab } from './MainAITab';
+export { default as MapDataFullTab } from './MapDataFullTab';
+export { default as MapLocationAITab } from './MapLocationAITab';
+export { default as MiscStateTab } from './MiscStateTab';
+export { default as ThemeHistoryTab } from './ThemeHistoryTab';
+export { default as TravelPathTab } from './TravelPathTab';
+export * from './tabUtils';
